refactor(main): extract _register_group2_services to backend/core/registrations/group2_services (audit A8, refs #153)
Commit 2 of N in PR A8.

Same mechanical pattern as commit 1: lift function body verbatim into
new module, leave 3-line shim in main.py. ~290 LOC moved.

  main.py: 2063 -> 1817 (-246 LOC)

Refs #145, refs #153
